update of 14/04/2022

+++ app/src/Admin/employee.php
@@ -65,7 +65,11 @@ class Employee {
     public static function updateEmployee($id, $name, $surname, $role, $username, $password, $admin){
         require_once(realpath(dirname(__FILE__, 3)) . '/vendor/autoload.php');
         $db = Database::getDB();
-        $sql = 'UPDATE Dipendente SET Nome = ?, Cognome = ?, Ruolo = ?, Username = ?, Password = ?, UserType = ? WHERE (id = ?)';
+        if (empty($password)){
+            $sql = 'UPDATE Dipendente SET Nome = ?, Cognome = ?, Ruolo = ?, Username = ?, UserType = ? WHERE (id = ?)';
+        } else {
+            $sql = 'UPDATE Dipendente SET Nome = ?, Cognome = ?, Ruolo = ?, Username = ?, Password = ?, UserType = ? WHERE (id = ?)';
+        }
         $stmt = $db->prepare($sql);
         if (!$stmt) {
             throw DatabaseException::queryPrepareFailed();
@@ -75,9 +79,15 @@ class Employee {
         } else {
             $admin = WORKER_USER;
         }
-        $password = password_hash($password, PASSWORD_DEFAULT);
-        if (!$stmt->bind_param('sssssii', $name, $surname, $role, $username, $password, $admin, $id)) {
-            throw DatabaseException::bindingParamsFailed();
+        if (empty($password)){
+            if (!$stmt->bind_param('ssssii', $name, $surname, $role, $username, $admin, $id)) {
+                throw DatabaseException::bindingParamsFailed();
+            }
+        } else {
+            $password = password_hash($password, PASSWORD_DEFAULT);
+            if (!$stmt->bind_param('sssssii', $name, $surname, $role, $username, $password, $admin, $id)) {
+                throw DatabaseException::bindingParamsFailed();
+            }
         }
         if ($stmt->execute()) {
             //Success
@@ -88,23 +98,25 @@ class Employee {
         return false;
     }
 
-    public static function deleteEmployee($id){
+    public static function deleteEmployee($employeId, $loggedId){
         require_once(realpath(dirname(__FILE__, 3)) . '/vendor/autoload.php');
-        $db = Database::getDB();
-        $sql = 'DELETE FROM Dipendente WHERE Dipendente.id = ?';
-        $stmt = $db->prepare($sql);
-        if (!$stmt) {
-            throw DatabaseException::queryPrepareFailed();
+        if ($loggedId != $employeId){
+            $db = Database::getDB();
+            $sql = 'DELETE FROM Dipendente WHERE Dipendente.id = ?';
+            $stmt = $db->prepare($sql);
+            if (!$stmt) {
+                throw DatabaseException::queryPrepareFailed();
+            }
+            if (!$stmt->bind_param('i', $employeId)) {
+                throw DatabaseException::bindingParamsFailed();
+            }
+            if ($stmt->execute()) {
+                //Success
+                return true;
+            } else {
+                throw DatabaseException::queryExecutionFailed();
+            }
         }
-        if (!$stmt->bind_param('i', $id)) {
-            throw DatabaseException::bindingParamsFailed();
-        }
-        if ($stmt->execute()) {
-            //Success
-            return true;
-        } else {
-            throw DatabaseException::queryExecutionFailed();
-        }
-        return false;
+        throw DatabaseException::cantDeleteCurrentUser();
     }
 }
